population-service/requirements.txt
grpcio>=1.43.0
grpcio-health-checking>=1.43.0
grpcio-reflection>=1.43.0
protobuf>=3.5.0.post1
pika>=1.1.0
qwikidata>=0.4.0
fluent-logger>=0.9.6
prometheus-client>=0.9.0
python-dotenv>=0.15.0
grpc-interceptor>=0.14.1